pyproject.toml
[project]
name = "webnect"
version = "1.0.0"
description = "A web interface for Kinect v1 using FastAPI"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "freenect>=0.1.0",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "uvicorn>=0.35.0",
]
